EXTREMELY HACKY SAPLING SUPPORT.
This is not upstreamable code. This is a way to be able to run basic pre-commit hooks with Sapling.

It works, but it's not great, and it basically comments out a bunch of code.

Unfortunately due to the nature of modern pre-commit (that is a generic git hook manager, *not* a pre-commit checker), the options are either architecting an abstraction or hard-forking.